refactor: move custom user details into user package
