+++ src/main/java/com/jpcchaves/adotar/application/service/user/CustomUserDetailsService.java
@@ -1,0 +1,28 @@
+package com.jpcchaves.adotar.application.service.user;
+
+import com.jpcchaves.adotar.domain.exception.ResourceNotFoundException;
+import com.jpcchaves.adotar.infra.repository.UserRepository;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+@Service
+public class CustomUserDetailsService implements UserDetailsService {
+    private final UserRepository userRepository;
+
+    public CustomUserDetailsService(UserRepository userRepository) {
+        this.userRepository = userRepository;
+    }
+
+    @Override
+    public UserDetails loadUserByUsername(String usernameOrEmail)
+            throws UsernameNotFoundException {
+        return userRepository
+                .findByUsernameOrEmail(usernameOrEmail, usernameOrEmail)
+                .orElseThrow(
+                        () ->
+                                new ResourceNotFoundException(
+                                        "Usuário não encontrado com os dados informados!"));
+    }
+}
